Restructure pants example project
