+++ ideprobe/tests/src/test/resources/OpenProjectTestBsp.conf
@@ -1,17 +1,27 @@
 include required("OpenProjectTest.conf")
 
-pants.import.targets = ["java_app::"]
+pants.import.targets = ["core/src/main/java::"]
 
 project {
-  name = "java_app"
+  name = "core.src.main.java"
   modules = [
     {
-      name = "java_app:main-bin",
-      contentRoots = [
+      name = "core_src_main_java:main-bin",
+      sourceRoots = [
         {
-          path = "java_app/src/com",
+          path = "core/src/main/java/com/example",
+          kind = "sources"
+          packagePrefix = "com.example"
+        }
+      ]
+    },
+    {
+      name = "core_src_main_java_com_example_sub:sub",
+      sourceRoots = [
+        {
+          path = "core/src/main/java/com/example/sub",
           kind = "sources",
-          packagePrefix = "src.com" // TODO this is a bug, it should be just "com"
+          packagePrefix = "com.example.sub"
         }
       ]
     }
