+++ ideprobe/tests/src/test/resources/OpenProjectTestPants.conf
@@ -1,22 +1,37 @@
 include required("OpenProjectTest.conf")
 
-pants.import.targets = ["java_app::", "python_bin::"]
+pants.import.targets = ["core/src/main::"]
 
 project {
-  name = "java_app::__python_bin::"
+  name = "core.src.main::"
   modules = [
     {
-      name = "java_app_main-bin",
-      contentRoots = [
-        { path = "java_app", kind = "sources" }
+      name = "core_src_main_java_main-bin",
+      sourceRoots = [
+        {
+          path = "core/src/main/java/com/example",
+          kind = "sources",
+          packagePrefix = "com.example"
+        }
       ]
     },
     {
-      name = "python_src", // is this a bug? the directory name is python_bin
-      contentRoots = [
-        { path = "python_bin", kind = "sources" }
+      name = "core_src_main_java_com_example_sub_sub",
+      sourceRoots = [
+        {
+          path = "core/src/main/java/com/example/sub",
+          kind = "sources",
+          packagePrefix = "com.example.sub"
+        }
+      ]
+    },
+    {
+      name = "python_src",
+      sourceRoots = [
+        { path = "core/src/main/python", kind = "sources" }
       ]
     }
   ],
   pythonModules = ["python_src"]
+  buildFile = "core/src/main/java/BUILD"
 }
