+++ ideprobe/tests/src/test/resources/PantsSettingsTest.conf
@@ -1,3 +1,3 @@
 include required("projects/PantsExample/ideprobe.conf")
 
-targetPath = "java_app"
+targetPath = "core/src/main/java"
