+++ ideprobe/tests/src/test/resources/RunTestsTest.conf
@@ -1,8 +1,14 @@
 include required("projects/PantsExample/ideprobe.conf")
 
-pants.import.targets = ["scala_tests"]
+pants.import.targets = ["core/src/test/scala::"]
 
-modulesWithTests {
-  bsp = ["scala_tests:scala_tests"]
-  pants = ["scala_tests_scala_tests"]
+runConfiguration {
+  bsp {
+    module.name = "core_src_test_scala:scala",
+    packageName = "com.example"
+  }
+  pants {
+    module.name = "core_src_test_scala_scala",
+    packageName = "com.example"
+  }
 }
