+++ ideprobe/tests/src/test/resources/projects/PantsExample/core/src/main/java/BUILD
@@ -1,0 +1,5 @@
+jvm_binary(name = 'main-bin',
+           sources = ['com/example/HelloWorld.java'],
+           main = 'com.example.HelloWorld',
+           basename = 'hello-example',
+           )
