+++ ideprobe/tests/src/test/resources/projects/PantsExample/core/src/main/java/com/example/HelloWorld.java
@@ -1,8 +1,8 @@
-package com;
+package com.example;
 
 public class HelloWorld {
     public static void main(String[] args) {
-	System.out.println("Hello, world!");
+        System.out.println("Hello, world!");
     }
 }
