+++ ideprobe/tests/src/test/resources/projects/PantsExample/core/src/main/java/com/example/sub/BUILD
@@ -1,0 +1,1 @@
+java_library(sources = ['*.java'])
