+++ ideprobe/tests/src/test/resources/projects/PantsExample/core/src/main/java/com/example/sub/Sub.java
@@ -1,0 +1,5 @@
+package com.example.sub;
+
+public class Sub {
+
+}
