+++ ideprobe/tests/src/test/resources/projects/PantsExample/core/src/main/python/BUILD
@@ -1,0 +1,3 @@
+python_binary(
+    source = "main.py"
+)
