+++ ideprobe/tests/src/test/resources/projects/PantsExample/core/src/main/python/main.py
@@ -1,0 +1,5 @@
+def func():
+    return ''
+
+if __name__ == '__main__':
+    func()
