+++ ideprobe/tests/src/test/resources/projects/PantsExample/core/src/test/scala/BUILD
@@ -1,5 +1,5 @@
 junit_tests(
-    sources = ['ExampleTests.scala'],
+    sources = ['com/example/ExampleTests.scala'],
     dependencies=[
         '3rdparty:junit',
         '3rdparty:scalatest',
