+++ ideprobe/tests/src/test/resources/projects/PantsExample/core/src/test/scala/com/example/ExampleTests.scala
@@ -1,3 +1,5 @@
+package com.example
+
 import org.junit.runner.RunWith
 import org.scalatest.WordSpec
 import org.scalatest.junit.JUnitRunner
