+++ ideprobe/tests/src/test/resources/projects/PantsExample/ideprobe.conf
@@ -7,4 +7,4 @@ probe {
   workspace.path = "classpath:/projects/PantsExample"
 }
 
-pants.import.targets = ["java_app::"]
+pants.import.targets = ["core::"]
