--- ideprobe/tests/src/test/resources/projects/PantsExample/java_app/BUILD
@@ -1,6 +1,0 @@
-jvm_binary(name = 'main-bin',
-  sources = ['src/**/*.java'],
-  main = 'com.HelloWorld',
-  basename = 'hello-example',
-)
-
