--- ideprobe/tests/src/test/resources/projects/PantsExample/python_bin/BUILD
@@ -1,3 +1,0 @@
-python_binary(
-    source = "pyexample.py"
-)
