--- ideprobe/tests/src/test/resources/projects/PantsExample/python_bin/pyexample.py
@@ -1,5 +1,0 @@
-def dump():
-    return 'dump'
-
-if __name__ == '__main__':
-    dump()
