+++ ideprobe/tests/src/test/resources/projects/ThriftExample/java-app/BUILD
@@ -1,4 +1,4 @@
 java_library(name='main',
-  sources=['Main.java',],
+  sources=['Main.java'],
   dependencies=['3rdParty:thrift', 'thrift-lib:thrift-lib']
 )
